requirements.txt
# PPT Master — Python runtime dependencies
#
# Target Python: 3.10+  (3.14.x recommended — matches brew-maintained latest stable)
# Install (core, required for finalize → pptx export):
#     cd ppt-master && pip3 install --user --break-system-packages -r requirements.txt
#
# For brew-managed Python, --user and --break-system-packages are REQUIRED because
# brew enables PEP 668 externally-managed-environment protection on site-packages.
#
# If you only need a subset of the pipeline, install the groups you need:
#   - CORE       : always required (validate, finalize, native pptx export)
#   - RENDER     : required by the "compat" svg→pptx rendering path (SVG→PNG raster)
#   - SOURCE2MD  : required when converting user-provided PDF / DOCX / EPUB / notebooks
#   - FETCH      : required when scraping WeChat / high-security URLs in web_to_md
#   - IMAGEGEN   : required by AI image_gen backends (OpenAI / Google Gemini)

# ==========================
# CORE — mandatory
# ==========================
Pillow>=10.0            # PIL/Image — used by finalize_svg + crop + embed
numpy>=1.24             # used by svg_finalize color math
python-pptx>=1.0        # PPTX authoring (native DrawingML mode)
lxml>=4.9               # python-pptx peer dep; also used directly
requests>=2.28          # HTTP client, image download, palette fetch
urllib3>=1.26           # pinned via requests; explicit for clarity
XlsxWriter>=3.0         # pptx compat dependency

# ==========================
# RENDER — required for the --compat raster fallback in svg_to_pptx
# If you only ever export Native (editable) PPTX, these are optional.
# ==========================
cairosvg>=2.7
svglib>=1.5
reportlab>=4.0

# ==========================
# SOURCE2MD — required when converting source documents
# ==========================
beautifulsoup4>=4.12    # imported as bs4 in web_to_md / doc_to_md
mammoth>=1.6            # DOCX → markdown
EbookLib>=0.18          # imported as ebooklib (EPUB)
PyMuPDF>=1.23           # imported as fitz (PDF)
markdownify>=0.11
nbconvert>=7.0          # .ipynb → markdown
nbformat>=5.9

# ==========================
# FETCH — required for web_to_md on TLS-guarded sites (WeChat etc.)
# ==========================
curl-cffi>=0.6

# ==========================
# IMAGEGEN — required only when calling the corresponding AI image backend
# Install the one you actually use; the others can be omitted.
# ==========================
openai>=1.0             # backend_openai
google-genai>=0.3       # imported as google.genai — Gemini image backend
